Refactor Planner Page component and add Navegante section
- Replace Carousel with new Navegante and Partners components
- Update i18n translations for new Planner Page structure
- Add styles for trip planner container
- Update tsconfig.json to use "preserve" for JSX
