+++ frontend/components/planner/Navegante/styles.module.css
@@ -1,0 +1,16 @@
+.container {
+    background-color: var(--color-primary-black);
+
+    & h1 {
+        color: var(--color-brand);
+    }
+
+    & h6 {
+        color: var(--color-primary-white);
+    }
+}
+
+.button {
+    background-color: var(--color-primary-white);
+    color: var(--color-primary-black);
+}
