+++ frontend/components/planner/Partners/styles.module.css
